Administrace Zařízení  (#20)
* zařízení

* zařízení

* přidání grafů, úprava chyba, administrace zařízení

* administrace zařízení

+++ backend/api/auth/delete.js
@@ -1,7 +1,10 @@
 const router = require('express').Router()
 const auth = require('../../modules/auth')
+let { DB } = require('mongquick')
+const db = new DB(process.env.MongoLogin)
 
 router.post('/', async (req, res) => {
+	try {
 	res.header('Content-Type', 'application/json')
 	if (!auth.isOriginAllowed(req.get('origin')))
 		return res.status(401).send({
@@ -13,9 +16,17 @@ router.post('/', async (req, res) => {
 		return res
 			.status(409)
 			.send({ valid: false, response: 'Uživatel s tímto jménem nexistuje' })
-
+if (!auth.authenticatePassword(req.body.password, await db.get(req.body.username).user.password)) return res.status(401).send({ valid: false, response: 'heslo není správné' })
 	auth.deleteUser(req.body.username)
 	res.status(200).send({ valid: true })
+}
+catch (err) {
+	console.log(err)
+	res.status(500).send({
+		valid: false,
+		response: 'Nastala chyba, ' + err
+	})
+}
 })
 
 module.exports = router
